btrfs: reduce size of struct btrfs_block_rsv
The full and failfast members are just bool indicators, type fits to a
few bits, we can reduce the overall structure size to 32 bytes by using
smaller types. Note that bitfields are not a good idea here as the
members are preceded by a spinlock and this has bad consequences
(non-atomic RMW updates to 8B word and spinlock).

As btrfs_block_rsv is embedded in several structures, their size also
shrinks (btrfs_fs_info -48B, btrfs_pending_snapshot -8B).

+++ fs/btrfs/ctree.h
@@ -454,9 +454,9 @@ struct btrfs_block_rsv {
 	u64 reserved;
 	struct btrfs_space_info *space_info;
 	spinlock_t lock;
-	unsigned short full;
-	unsigned short type;
-	unsigned short failfast;
+	u8 full;
+	u8 type;
+	u8 failfast;
 };
 
 /*
